requirements.txt
streamlit
pdfplumber
tiktoken
openai
faiss-cpu
